Remove tags column from Posts table
Initially, I presumed it would be smart to submit the tags array
directly from the new post form @ posts#new, but when I tried to
write the form and submit an array, I found out it wasn't simple.
Tried some research on how to do it, and found out that I would
need to use a lot of stuff completely out of my current knowledge
to do it through the form.

Also in my research, I found two possible solutions to implement
post tags in my blog: act-as-taggable-on and simply create a Tags
entity in my Rails app. Since these seems more simple and easy to
get my blog working, the :tags column at Posts table is no longer
necessary, so this migration removes the table from the schema.

Also, I'm testing a new way to submit commit messages with more
context to my repos.