Add Devise to the application
To handle user sessions and enable Omniauth in the application,
I'll use the Devise gem in this project. There's no need to reinvent
the wheel of user authentication, or struggle to implement OAuth2.

For now, I have only set up the development environment variables for
Devise. Later in this project I'll figure it out how to enable it for
production.