Add User model through Devise
For now, to keep the project (and the commits) simple, we'll start
with the default model generated by Devise. Since we will yet
implement an Admin role for the user, then adjust the posts views
to only allow the Admin to create new posts and all users to
comment on them, the main focus now is to assure it works properly.