Add navigation links to top of application layout
Add the partial _navigation.html.erb to application layout, which
will become the foundation to the blog's navigation bar. For now,
it will hold the links regarding the user's session and registration
(sign in, sign up, log out and edit profile).

Also, the navigation has the "Eliseu on Rails" header, so I decided
to remove the header from the posts#index view to avoid a duplicate.