Add navigation bar style customization
Add Tailwind classes to style the navigation bar. Found out about
the button_to command, but it sets a form with a class of button_to
which prevented using some flexbox properties which I wanted to use
in the navigations buttons.

It's a first design, and designing isn't my forte, so I'm going
for a minimal-I'm-not-a-designer look for the blog, until I get
some help to think a better design.